GRAPHICS: Move common OpenGL files into graphics/opengl
