+++ backends/platform/android/android.h
@@ -31,8 +31,8 @@
 #include "graphics/palette.h"
 #include "graphics/surface.h"
 #include "graphics/pixelbuffer.h"
-#include "graphics/opengles2/system_headers.h"
-#include "graphics/opengles2/framebuffer.h"
+#include "graphics/opengl/system_headers.h"
+#include "graphics/opengl/framebuffer.h"
 #include "backends/base-backend.h"
 #include "backends/plugins/posix/posix-provider.h"
 #include "backends/fs/posix/posix-fs-factory.h"
