+++ graphics/module.mk
@@ -25,10 +25,10 @@ MODULE_OBJS := \
 	yuv_to_rgb.o \
 	yuva_to_rgba.o \
 	pixelbuffer.o \
-	opengles2/extensions.o \
+	opengl/framebuffer.o \
+	opengl/texture.o \
+	opengl/extensions.o \
 	opengles2/shader.o \
-	opengles2/framebuffer.o \
-	opengles2/texture.o \
 	opengles2/box_shaders.o \
 	opengles2/control_shaders.o \
 	opengles2/compat_shaders.o \
